Fixed delete playlist button error
When the "delete" button was clicked in EditForm without a playlist
actually selected, an exception was thrown. The value of
lsbx_EditListBox.SelectedIndex was equal to -1 when nothing was actually
selected. So now before trying to delete the data in that index, it
first checks to see if an index is even selected. Updated comments.