Fixed SaveTextFile error
When PandQ.saveTextFile is executed, it saves to the file path that
saved as PandQ.file. When PandQ.file = null, an exception is thrown.
Now, in saveTextFile is executed, it first checks to see if there is a
file at all. If not, the user chooses a place to save.

in ViewAllQuestionsForm, it checks to see if any questions were even
selected, removing the error.